# docs/DOC-05-api-spec.yaml
openapi: 3.1.0
info:
  title: Contract Intelligence API
  version: 0.3.0
  status: enterprise-ready — đã chốt kiến trúc Bearer JWT, x-rbac & Tenant Isolation (Sprint 1)
  description: |
    REST API cho Contract Intelligence — hệ thống trích xuất, so sánh và review hợp đồng cấp Enterprise.

    **Khớp kiến trúc:** `docs/DOC-04-architecture.md` (v0.1, 2026-09-16)
    **Khớp schema:** `docs/DOC-04b-postgres-schema.sql` (PostgreSQL v1.0 đã chốt 2026-09-17)

    ## Cơ chế Xác thực & Phân quyền Enterprise
    - **Xác thực (Bearer JWT):** Mọi request (trừ probe `/healthz`, `/readyz`) bắt buộc mang header:
      `Authorization: Bearer <jwt_token>`
      JWT payload mang các claims: `sub` (user_id), `tenant_id`, `role` (`OPERATOR` | `REVIEWER` | `ADMINISTRATOR`).
    - **Tenant Isolation (`X-Tenant-Id`):** Client bắt buộc gửi header `X-Tenant-Id`.
      Backend enforce tenant boundary nghiêm ngặt: `X-Tenant-Id` phải trùng khớp với claim `tenant_id` trong JWT. Mọi thao tác ghi/đọc dữ liệu đều bị cô lập theo tenant.
    - **Phân quyền Role-Based Access Control (`x-rbac`):** Mỗi endpoint quy định chặt chẽ danh sách role hợp lệ qua extension property `x-rbac`:
      - `OPERATOR`: Khởi tạo dossier/batch, upload văn bản, xác nhận manifest, kích hoạt run, tạo yêu cầu Re-OCR.
      - `REVIEWER`: Xem dossier/documents/facts/conflicts/citations, thực hiện review actions, xem lịch sử revision, đề xuất khóa dossier.
...

    ## Đơn vị xử lý & Nghiệp vụ cốt lõi (9 Phân hệ Enterprise)
...

...

...

  contact:
...
servers:
...
paths:
...
  /dossiers/{dossier_id}/review-items:
    parameters:
      - $ref: '#/components/parameters/DossierId'
      - name: status
        in: query
        schema:
          type: string
          enum: [open, awaiting_evidence, resolved]
      - name: priority
        in: query
        schema: { type: string, enum: [P1, P2, P3] }
    get:
      tags: [review]
      summary: Hàng đợi review
      operationId: listReviewItems
      x-rbac: [REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Review item list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeReviewItemList'
  /review-items/{review_item_id}:
    parameters:
      - name: review_item_id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [review]
      summary: Chi tiết review item (kèm version & target snapshot)
      operationId: getReviewItem
      x-rbac: [REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Review item detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeReviewItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /review-items/{review_item_id}/actions:
    parameters:
      - name: review_item_id
        in: path
        required: true
        schema: { type: string }
      - name: Idempotency-Key
        in: header
        required: false
        schema: { type: string, format: uuid }
    post:
      tags: [review]
      summary: Submit review action (Confirm / Correct / Reject / Needs more evidence)
      description: |
        **Optimistic concurrency bắt buộc:**
        - Client gửi `base_version` (là `current_version` đã đọc trước đó, hoặc `0` nếu dựa trên máy).
        - Nếu DB hiện đang ở version khác → API trả **`409 Conflict`** kèm state mới nhất để client refetch.
        - Trong vòng 60s sau lần submit đầu, request lặp lại với cùng `Idempotency-Key` sẽ trả cùng kết quả (retry an toàn).
        - Với `correct`, `corrected_value` hoặc `corrected_bbox` phải có (CHECK constraint).
      operationId: submitReviewAction
      x-rbac: [REVIEWER, ADMINISTRATOR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewActionRequest'
      responses:
        '200':
          description: Action recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeReviewAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Revision conflict — concurrent update detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeReviewConflict'
  /review-items/{review_item_id}/revisions:
    parameters:
      - name: review_item_id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [revisions]
      summary: Lấy chuỗi lịch sử phiên bản thẩm tra bất biến (append-only review revisions)
      description: |
        Tuân thủ quy chuẩn kiểm toán Enterprise: Mọi thao tác review (confirm, correct, reject)
        đều tạo một revision bất biến mới (append-only) chứ không ghi đè, kèm diff, author và version.
      operationId: listReviewItemRevisions
      x-rbac: [REVIEWER, ADMINISTRATOR]
      responses:
        '200':
          description: Append-only review revisions trail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelopeReviewRevisions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Approval ────────────────────────────────────────────────────────────
...
components:
...
  parameters:
...
    DossierId:
      name: dossier_id
      in: path
      required: true
      schema: { type: string, pattern: "^dos_[0-9A-HJKMNP-TV-Z]{26}$" }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: |
            Mã lỗi nội bộ (xem DOC-04 mục 8.6).
            Ví dụ: `PDF_INVALID`, `OCR_RATE_LIMITED`, `DB_ERROR`, `INTERNAL_ERROR`.
        detail: { type: string }
        instance: { type: string }
        trace_id:
          type: string
          description: ID trace OpenTelemetry để debug
...
    JobStatus:
      type: string
      enum: [uploaded, processing, extracted, pending_review, reviewed, approved, failed, cancelled]
      description: Trạng thái job (BRD §14); cancelled khi dossier bị xóa (tombstone)
...
    Disposition:
      type: string
      enum:
        - comparable_match
        - comparable_difference
        - candidate_amendment
        - not_comparable
        - insufficient_evidence
    Severity:
      type: string
      enum: [high, medium, low]
    FindingType:
      type: string
      enum: [structured, semantic]
    FindingScope:
      type: string
      enum: [within_document, contract_annex, annex_annex]
    ReviewActionType:
      type: string
      enum: [confirm, correct, reject, needs_more_evidence]
    ReviewItemStatus:
      type: string
      enum: [open, awaiting_evidence, resolved]
    ReviewItemPriority:
      type: string
      enum: [P1, P2, P3]
    TargetType:
      type: string
      enum: [fact, finding, annex_link, clause_node, table_cell, citation]

    # ── Geometry ───────────────────────────────────────────────────────────
    BBox:
      type: object
      description: Canonical Page Space (CPS) — `[x0,y0,x1,y1]` chuẩn hóa 0..1
      required: [x0, y0, x1, y1]
      properties:
        x0: { type: number, minimum: 0, maximum: 1 }
        y0: { type: number, minimum: 0, maximum: 1 }
        x1: { type: number, minimum: 0, maximum: 1 }
        y1: { type: number, minimum: 0, maximum: 1 }
    BBoxLevel:
      type: string
      enum: [word_union, line, clause_region]
    BBoxSource:
      type: string
      enum: [native, detector, estimated, human]
    CoordSystem:
      type: object
      required: [space, range, origin, format]
      properties:
        space: { type: string, enum: [CPS] }
        range: { type: string, enum: ["0..1"] }
        origin: { type: string, enum: ["top-left"] }
        format: { type: string, enum: ["x0,y0,x1,y1"] }

    # ── Dossier ────────────────────────────────────────────────────────────
...
    ClauseNode:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        node_type: { type: string, enum: [annex, article, clause, point, preamble, signature_block] }
        label: { type: string, description: 'VD: "Điều 5"' }
        number: { type: string, nullable: true }
        title: { type: string, nullable: true }
        text: { type: string }
        lang: { type: string, nullable: true }
        parent_id: { type: string, nullable: true }
        stable_path: { type: string, description: 'VD: "annex-01/art-5/cl-2/pt-a"' }
        page_start: { type: integer }
        page_end: { type: integer }
        confidence: { type: number }
        regions:
          type: array
          items:
            type: object
            properties:
              page_no: { type: integer }
              bbox: { $ref: '#/components/schemas/BBox' }
              bbox_source: { $ref: '#/components/schemas/BBoxSource' }
        children:
          type: array
          items: { $ref: '#/components/schemas/ClauseNode' }

    # ── Tables ─────────────────────────────────────────────────────────────
...
    CitationSegment:
      type: object
      required: [page_no, char_start, char_end, bbox]
      properties:
        page_no: { type: integer }
        line_id: { type: string }
        char_start: { type: integer }
        char_end: { type: integer }
        bbox: { $ref: '#/components/schemas/BBox' }
        bbox_level: { $ref: '#/components/schemas/BBoxLevel' }
    Citation:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        run_id: { type: string }
        quote: { type: string }
        quote_sha256: { type: string }
        segments:
          type: array
          items: { $ref: '#/components/schemas/CitationSegment' }
        doc_char_span:
          type: array
          items: { type: integer }
          minItems: 2
          maxItems: 2
        coord_system: { $ref: '#/components/schemas/CoordSystem' }
...
    Fact:
      type: object
      properties:
        id: { type: string }
        document_id: { type: string }
        run_id: { type: string }
        key: { type: string, description: 'VD: "price.total", "party.name"' }
        fact_type: { type: string }
        raw_text: { type: string }
        normalized_value: {}
        context_clause_id: { type: string, nullable: true }
        context_text: { type: string, nullable: true }
        confidence: { type: number }
        extractor: { type: string }
        validation_status: { type: string, enum: [passed, failed, skipped] }
        citation: { $ref: '#/components/schemas/Citation' }
        trace_id: { type: string, nullable: true }
    FactEffective:
      type: object
      description: Fact kèm giá trị hiệu lực sau review + ETag cho optimistic concurrency
      properties:
        fact: { $ref: '#/components/schemas/Fact' }
        machine_value: {}
        effective_value: { nullable: true }
        review_state:
          type: string
          enum: [unreviewed, confirmed, corrected, rejected, awaiting_evidence]
        reviewer_id: { type: string, nullable: true }
        reviewed_at: { type: string, format: date-time, nullable: true }
        review_item_id:
          type: string
          nullable: true
          description: ID của review_item tương ứng (NULL = chưa sinh item)
        current_version:
          type: integer
          minimum: 0
          description: |
            Version hiện tại của review_item (0 = giá trị máy, 1..N = sau N lần action).
            **Client PHẢI gửi lại giá trị này trong `base_version` khi submit action.**

    # ── Finding ────────────────────────────────────────────────────────────
    FindingSide:
      type: object
      properties:
        side: { type: string, enum: [a, b] }
        document_id: { type: string }
        document_role: { type: string, enum: [contract, annex] }
        fact_id: { type: string, nullable: true }
        clause_node_id: { type: string, nullable: true }
        citation: { $ref: '#/components/schemas/Citation' }
        value_snapshot: { nullable: true }
    Finding:
      type: object
      properties:
        id: { type: string }
        dossier_id: { type: string }
        run_id: { type: string }
        finding_type: { $ref: '#/components/schemas/FindingType' }
        scope: { $ref: '#/components/schemas/FindingScope' }
        key_or_topic: { type: string }
        disposition: { $ref: '#/components/schemas/Disposition' }
        severity: { $ref: '#/components/schemas/Severity' }
        confidence: { type: number }
        rationale: { type: string }
        method: { type: string }
        sides:
          type: array
          minItems: 2
          maxItems: 2
          items: { $ref: '#/components/schemas/FindingSide' }
        disclaimer:
          type: string
          default: "Kết quả so sánh kỹ thuật, không phải kết luận pháp lý."
        review:
          type: object
          nullable: true
          properties:
            item_id: { type: string }
            status: { $ref: '#/components/schemas/ReviewItemStatus' }
            current_version: { type: integer }

    # ── Review item & action ───────────────────────────────────────────────
    ReviewItem:
      type: object
      properties:
        id: { type: string }
        dossier_id: { type: string }
        run_id: { type: string }
        target_type: { $ref: '#/components/schemas/TargetType' }
        target_id: { type: string }
        reason: { type: string }
        priority: { $ref: '#/components/schemas/ReviewItemPriority' }
        status: { $ref: '#/components/schemas/ReviewItemStatus' }
        version:
          type: integer
          minimum: 1
          description: Version hiện tại của item — client gửi lại trong `base_version` khi submit action
        source_trace_id: { type: string, nullable: true }
        source_observation_id: { type: string, nullable: true }
        target_snapshot:
          description: Snapshot target để reviewer thấy ngay không cần join
          oneOf:
            - $ref: '#/components/schemas/FactEffective'
            - $ref: '#/components/schemas/Finding'
            - $ref: '#/components/schemas/ClauseNode'
        created_at: { type: string, format: date-time }
    ReviewActionRequest:
      type: object
      required: [action, base_version]
      properties:
        action: { $ref: '#/components/schemas/ReviewActionType' }
        base_version:
          type: integer
          minimum: 0
          description: |
            Version hiện tại của review_item mà client đang thấy. **0** = dựa trên máy.
            **1, 2, ...** = sau N lần action.
            Mismatch với DB → 409 Conflict.
        corrected_value:
          description: Bắt buộc nếu action=correct và không có corrected_bbox
        corrected_bbox:
          type: array
          description: Bắt buộc nếu action=correct và không có corrected_value (cho citation)
          items:
            type: object
            properties:
              page_no: { type: integer }
              bbox: { $ref: '#/components/schemas/BBox' }
        comment: { type: string, maxLength: 2000 }
    ReviewActionResponse:
      type: object
      properties:
        review_action_id: { type: string }
        item_status: { $ref: '#/components/schemas/ReviewItemStatus' }
        new_version: { type: integer }
        effective_value: { nullable: true }
        machine_value: { nullable: true }
        job_status: { $ref: '#/components/schemas/JobStatus' }
        open_items_remaining: { type: integer }
        idempotent_replay:
          type: boolean
          description: True nếu response này từ Idempotency-Key replay (60s window)
    ReviewConflictResponse:
      type: object
      description: Trả về khi `base_version` không khớp DB
      properties:
        error:
          type: object
          properties:
            code: { type: string, enum: [VERSION_CONFLICT] }
            message: { type: string, example: "Version conflict: another action was submitted before yours (current version: 4)." }
        current_state:
          description: Trạng thái hiện tại của target để client refetch và reconcile
          oneOf:
            - $ref: '#/components/schemas/FactEffective'
            - $ref: '#/components/schemas/Finding'
        your_submitted_action:
          type: object
          nullable: true
          description: Nếu action của bạn vẫn được ghi nhận dù version mismatch (server policy), thông tin action ở đây
...
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          nullable: true
        meta:
          type: object
          properties:
            request_id: { type: string }
            trace_id: { type: string }
            timestamp: { type: string, format: date-time }
    Pagination:
      type: object
      properties:
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
        total:
          type: integer
          nullable: true
          description: Có thể null nếu không đếm được (cursor-based)
...
    ApiEnvelopeReviewItem:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ReviewItem' }
    ApiEnvelopeReviewItemList:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ReviewItem' }
            pagination: { $ref: '#/components/schemas/Pagination' }
    ApiEnvelopeReviewAction:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ReviewActionResponse' }
    ApiEnvelopeReviewConflict:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ReviewConflictResponse' }
...
    ReviewItemRevision:
      type: object
      required: [revision_number, action, author_user_id, created_at]
      properties:
        revision_number: { type: integer }
        action:
          type: string
          enum: [confirm, correct, reject, needs_more_evidence]
        author_user_id: { type: string }
        author_role: { type: string }
        comment: { type: string, nullable: true }
        corrected_value: { type: object, nullable: true }
        corrected_bbox:
          type: array
          items: { type: number }
          nullable: true
        previous_version: { type: integer }
        created_at: { type: string, format: date-time }

    # ── Optimization Schemas ────────────────────────────────────────────────
...
    ApiEnvelopeReviewRevisions:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ReviewItemRevision' }
...
  responses:
    Unauthorized:
      description: Token JWT không hợp lệ, đã hết hạn hoặc không được cung cấp (401)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Không có quyền truy cập theo vai trò x-rbac hoặc không khớp X-Tenant-Id (403)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    BadRequest:
      description: Request không hợp lệ (validation, schema)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token chuẩn RFC 7519 chứa các claims:
        - `sub`: User ID (ULID)
        - `tenant_id`: Định danh Tenant (bắt buộc khớp với header X-Tenant-Id)
        - `role`: OPERATOR | REVIEWER | ADMINISTRATOR
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: |
        Header định danh Tenant phục vụ phân lập đa khách hàng (Tenant Isolation).
        Bắt buộc phải trùng khớp với claim `tenant_id` trong Bearer JWT.

# ═══════════════════════════════════════════════════════════════════════════
# SECURITY, WEBHOOKS, CALLBACKS
# ═══════════════════════════════════════════════════════════════════════════
security:
  - bearerAuth: []
    tenantHeader: []
...